account for any when binding to equal bound
Summary:
When any appears in lower bounds then union the equal bound with any.
E.g. in
```
any, void <: #1 <: void
```
Previously we would solve to void, and lose the fact that any was involved. We want to solve to `_ | void` here.

This is a bit ad hoc, but I hope to supersede that very soon with D16457968. This is to unblock Pierre's work on global inference (see also added test).

Reviewed By: andrewjkennedy

Differential Revision: D16579907

fbshipit-source-id: 60893e311222923abab566c9e34432bc63a2c2b8