Use Hh_bucket in [next] and deprecate [inputs]
Summary:
As per Julia's suggestion in D16570870, use `Hh_bucket` to represent inputs. This accounts for the situation where not all workers have completed, and we want to keep workers alive until all work is **completed**, and not **pending**.

The logic to handle this isn't implemented yet (it'll be implemented later in the stack).

To summarize this diff, `next` has its signature changed to the following, and `inputs` has been deprecated.
```
type next = monitor_env -> monitor_env * 'input Hack_bucket.bucket
```

For more info, refer to D16605374 for all of the upcoming changes :)

Reviewed By: jewelpit

Differential Revision: D16610377

fbshipit-source-id: bca047af454e97f4d8297e6bd592b4ab08b8bd2f