Add initial_env and monitor_env
Summary:
To implement a memory cap, MultiWorker passes a `list ref` to `merge` and `next` in [typing_check_service](https://fburl.com/k14b5wff), to allow updating of the input list when files are unchecked.

Although the same approach is used in D16587649, it's probably better to make the signature of `Shared_lru.run` more general to allow the inputs to be updated by worker output.

The end goal of the next few diffs is the following:

```
val run :
  host_env: host_env -> // Value generated from Shared_lru.init, encodes file descriptors and channels
  initial_env: env -> // Initial environment, should encode the inputs and any other data
  job: (input -> output) -> // Job
  callback: (env -> output -> env) -> // Replaces [reduce], update the environment from the output
  next:(env -> env * Hack_bucket input) -> // Returns the next environment and the next input if there is one.
  'output
```

This diff lays some groundwork for the upcoming changes:
- Introduce `initial_env`.
- Rename `env` to `monitor_env`, and make `worker_map` a field of the record.

Reviewed By: jewelpit

Differential Revision: D16605374

fbshipit-source-id: fd98443b3b80c22b5e2e39027b25b44d330bdd5a