Add next function in Shared_lru.run
Summary:
The motivation behind this change is to ultimately implement a memory cap for workers. In the current implementation, `inputs: 'input list` and `job: 'input -> 'output`.

However, this means that currently `inputs` is fixed, and that jobs cannot affect `inputs` once it's passed to the `shared_lru` library. If a memory cap were implemented with the current architecture, there isn't any mechanism of communicating which files weren't typechecked, and modifying the initial input list.

To address this, this diff adds `next: 'inputs -> 'inputs * 'input option` to `Shared_lru.run` and `inputs: 'inputs`, which is a step towards making `inputs` dynamic.

Note that the type `'inputs` is not explicitly a `'input list`, although in all current uses it is.

Reviewed By: jewelpit

Differential Revision: D16570870

fbshipit-source-id: 767f974e1355dd0ec3a730b9f1e1a145586cdd27