rust_to_ocaml: perf improvement, use a dedicated field and not a hashtable for sharing
Summary:
An hashtable is used to implement sharing of tokens between Rust and OCaml.

Rather than copying tokens each time they are encountered (the parsetree is
actually a dag), we copy once and use an hashtable to memoize tokens.

However, this turns out to be a bottleneck. This patch use a dedicated mutable
field in Rust tokens representation to store a pointer to the OCaml value.

This is safe as long as the OCaml GC is not run / the Rust value is used within
the same calls to ocamlpool_enter/ocamlpool_leave.

Reviewed By: shiqicao

Differential Revision: D16541139

fbshipit-source-id: 77de7374f0da9502078bf93517676b74ee1ebd73