rust_to_ocaml, ocamlpool: a fast path for field initialization
Summary:
Until now, only caml_modify was used.
caml_initialize is faster and safer than caml_modify (the previous content is
not read, so the memory can contain garbage -- that's already the case,
previous patch disabled clearing of memory).

However this works only when initializing a value. Furthermore, we inline the
fast path in Rust:
- we assume that the values mutated were just allocated by OCamlpool
- the fast path is reached when most mutation points to OCamlpool or are
  integers.

TODO?: it should be correct and more general to test that the value is outside
the minor heap and not just inside the pool.  That might help when referencing
the linebreak map.

Reviewed By: shiqicao

Differential Revision: D16541138

fbshipit-source-id: c610ff3f85b27713fa0ccae0bab6d285d37059b7