Load naming table.
Summary: This change adds naming table saved state loading to hh_server.  We load the saved state in tandem with the regular one (mostly, we currently only step futures continuations when either `is_ready` or `get` are called), then add the list of changed files it returns to our list of files that need to be re-named.

Differential Revision: D16467284

fbshipit-source-id: 1b8073712f0b09e4d063679e2f259b904e2c87e7