Add notification emitter to `ClientIdeService.t`
Summary:
This is a stream of notifications that can be written to by the IDE service and read from by `clientLsp`. The notifications can be sent/received asynchronously, without interfering with synchronous RPC calls.

We'll use this in the immediate future to signal when the IDE service is finished processing changes for the saved state.

Reviewed By: jewelpit

Differential Revision: D16499502

fbshipit-source-id: 63a443333c4dabb76fd3d57618722bafc10f958b