Show detailed file context for unexpected notifications in LSP tests
Summary: Similarly to {D16612477}, we can show a bit of extra context when receiving an unexpected notification. We print the request that immediately preceded the notification as context. Note that there's no foolproof way to nail down which request was really associated with a given notification, but this should be enough to give a hint about what's going on.

Reviewed By: 2BitSalute

Differential Revision: D16632533

fbshipit-source-id: a2c82f4922c2dd0823670f5093f43f616d0afd37