Rearchitect sIDE to support notifications
Summary:
This is a big commit that rearchitects sIDE to support sending notifications at any time. Fortunately, I think it makes the architecture simpler than it was before, rather than more complex.

Before this commit, we could only issue pseudo-synchronous requests to the IDE daemon process. Only one request could be running at a time, and it would have to return before the next request could be sent. (We could, however, do other Lwt operations that didn't involve the IDE service in the meantime, although we don't currently do this.)

A "notification" is sent from the IDE daemon process to the main `clientLsp` process at any time, including while we're in the middle of an RPC call (that is, when we've sent a request and are waiting for a response). These notifications can be used to update the main `clientLsp` process about the status of the IDE service.

In this commit, we start emitting the only notification `Done_processing`, which is used to indicate that the queue of file changes to process has transitioned from non-empty to empty. This is useful 1) for logging purposes, to easily determine whether the IDE service has finished updating based on file changes, and 2) in tests so that we can confirm that we've waited for the IDE service to finish updating. We could introduce more notifications later.

I tried adding a callback that could be invoked in the middle of an RPC call to inform the caller about a notification, and also furnishing `serve` with such a callback. This ended up making the code too difficult to write correctly, since there were multiple places that notifications could be sent and received.

This commit changes it so that there's single sources of truth for various streams:

* The messages-to-send-to-the-IDE-daemon stream.
    * Previously, this stream was only present in the `Initialized` state. Now it's associated with all states, which makes it easier to get a handle to.
* The responses-received-from-the-IDE-daemon stream.
    * Previously, this wasn't a stream at all. Instead, we always assumed that the next message to be read was the response message. This no longer holds, since the next message can now be a notification.
    * Now that we have a stream, it also becomes possible in principle to issue requests in parallel and pair the responses to the requests at some time after receiving the responses, which would not have been easily written before.
* The notifications-received-from-the-IDE-daemon stream.
    * Previously, we had no such mechanism. Now, we write all notifications to this stream, and the caller can get a handle to the stream to start processing those notifications at their leisure.

In doing so, we've also made the IDE service less stateful. Previously, we had to issue a request to initialize the IDE service before it would even spawn a subprocess. Now, `make`ing the IDE service spawns the subprocess, and there isn't complicated state to carry along state transitions.

Reviewed By: jewelpit

Differential Revision: D16577167

fbshipit-source-id: e0c7a4d290fec7debd88b2a8ac8a2233ca14d8eb