Should not skip updating dependencies in check mode
Summary:
We've *always* skipped updating dependencies in check mode, but the original reason appears to have been that we intended the check mode to mean that there is no saved state.

I went back into the history of this line, and found this code in serverInit.ml from 2014 or even earlier:

```
  let is_check_mode =
    ServerArgs.check_mode genv.options &&
    ServerArgs.convert genv.options = None &&
    (* Note: we need to run update_files to get an accurate saved state *)
    ServerArgs.save_filename genv.options = None
  in

  if not is_check_mode then begin
    Typing_deps.update_files files_info;
  end;
```

Somewhere, in some refactoring, this original meaning must have been lost, and since we don't use --check for anything other than saving states and testing, I haven't been bothered by this behavior enough to figure out the root cause, until now.

Now, for saving states, we do end up updating the dependencies later on, but when running `hh_server --check` by itself, we end up not doing any actual type checking of the dependencies.

I would say that this makes the result of an `hh_server --check` quite incomplete.

Notice that I'm also removing the call to update the dependencies from places where saved states are not used: Eager init and full_init in Lazy init. If that call does anything, it's just wasted work.

Finally, I updated the name of the function and the verbiage in the logging to be more in line with what it actually does. It inherited its name from Typing_deps.update_file, but the name does not make sense in the context of an init module.

Differential Revision: D16142242

fbshipit-source-id: 653dee81f0684da9c99300c0140274aea20c24da