Always recompute bucket size in typing_check_service
Summary:
I noticed that the bucket size in `typing_check_service.ml` was computed prior to the closure being sent to the workers, which means that the bucket size would stay constant throughout typechecking.

This could introduce scenarios where we attempt to typecheck the same batch of files with a single worker over and over again, rather than distribute the work to more workers (with a smaller bucket size).

Ran a [perf test with 16 repetitions](https://fburl.com/everpaste/y9du1wgc), and it looks like there are performance wins.

Reviewed By: 2BitSalute, jewelpit

Differential Revision: D16632799

fbshipit-source-id: 4c1f53adfe866030b84d404e88907d37e10b9634