Flip sequential FAST extension <--> batch FAST extension
Summary:
We should not use batch FAST extension unless we're using a SQLite-backed naming table AND the number of files to add to FAST is greater than the threshold.

Here, we were falling through to batch extension even when not using SQLite.

"Extending FAST" means getting FileInfo from the naming table for each "additional" file and appending it to a list. In the full init case, all files are "additional", and in the full init case, we are never backed by SQLite.

Batch uses MultiWorker, sequential just does it in a loop. Using MultiWorker is faster for the SQLite-backed naming table.

In any case, on a full init, we are not backed by SQLite, so it's incorrect to follow this code path.

Reviewed By: kmeht

Differential Revision: D16655724

fbshipit-source-id: eba4dafcf11e1fe8deec84e14418e52ae52ccf14