Fix invariant with like types
Summary: `invariant` is rewritten to `invariant_violation` which is based on `printf`. The `printf` function takes an opaque type alias `HH\FormatString<T>`, which is pessimized to `~HH\FormatString<~T>`. This breaks the special-case handling in Typing_exts, which just looks at the type of the magic function. For a quick solution, I've added like types to the match cases.

Reviewed By: Matt-Schellhas

Differential Revision: D16521513

fbshipit-source-id: a3ec5e1f6d108fe204822f3a7d1aaaa1d7b2f760