Track a MixedArray keyTypes bitset
Summary:
For MixedArray and DictArray, there's a free byte in the 8-byte HeapObject header. In this diff, we make use of this byte to track a bit-set of key types in each array. Including them in this position is convenient, because we can (for example) simultaneously guard on an array being mixed and having certain types of keys.

There are four valid key types for these arrays: non-static strings, static strings, ints, and tombstones. Every time we update a key, we simply or the new key type into the bit-set. For most bulk operations, like NewStructDArray, CopyMixed, and compact, we can summarize the key effect of the operations with a single bit-set update.

To check the correctness of these methods, I've update checkInvariants. Note that we don't check key bit-set equality here, only that the set of true key types is a subset of the set of tracked keys. That's because this bit-set tracking can't distinguish between setting and erasing a string key (keyTypes = kTombstoneKey) and setting two string keys and erasing one (keyTypes = kStaticStrKey | kTombstoneKey).

On its own, this change is a small instructions and stores regression. I've included a small stores improvement (don't track static string keys, the common case), but the real wins come from diffs that build on this one. There are two main ideas:
1. Skip inc/dec checks for keys of arrays with only tombstone / int / static str keys.
2. Optimize static string lookups in arrays with static string keys.

Reviewed By: mofarrell

Differential Revision: D16536206

fbshipit-source-id: 8c328c825bf6c179b9e3e3ba746e279e9df8fa7d