Scan small MixedArrays with static keys
Summary:
Building on the MixedArray keyTypes bit-set tracking, this diff optimizes MixedArray lookups.

Profiling shows that there are many MixedArray lookups of a static string key, in a small array of static keys that don't get caught by our offset profiling. This situation comes up if a) there are optional keys in the array which throw the offset off or b) the key is typically missing. b) is the much more common case; it's specific to the ArrayIdx and DictIdx opcodes, because these ops don't throw on missing keys.

We update the ArrayAccessProfile to count cases where:
1. the array has capacity 3
2. the array has all static keys

If this situation happens more than 80% of the time, and if the offset profile didn't work, we pass a "small" size hint to irlower. The two irlower methods that use the "small" size hint check that the key is static before using scan-lookups instead of regular hash lookups.

Reviewed By: mofarrell

Differential Revision: D16539051

fbshipit-source-id: 7b07cde48e74c6e5e7913e18263797f8a63c5b9c