Consolidate reified hint checking in Reified_check
Summary: Previously, we only banned PHP arrays in type constants for the purpose of reification, but this was incorrect. To be reified, a type constant may not contain any of the types that cannot be reified. This was made clear after renaming `__DisallowPHPArrays` to `__Reifiable`. Now, when Type_const_check encounters that attribute, it defers to Reified_check instead of vice versa.

Reviewed By: oulgen

Differential Revision: D16665000

fbshipit-source-id: ba755025a9c1b667da7a1e5b31d70fcf8a529cf2