Don't send response messages to notifications
Summary:
When updating the serverless IDE to support notifications, I neglected to update one important site, which wasn't caught statically since it was at the RPC boundary.

Before this commit, the daemon process would always send a response to every message it received, including notifications messages -- which don't expect a response. This caused bizarre behavior where issuing certain IDE requests after you saved a file would generate the right results on the daemon side, but null-looking results on the client side, as the value `()` would be in the RPC queue as the next response.

Reviewed By: jewelpit

Differential Revision: D16663597

fbshipit-source-id: 23c81b8d7dc1886010d0b721748993684e10bf03