Fix bug with rendering `powered_by` in LSP test autofixes
Summary:
When an LSP test fails because a request was met with an incorrect response, we show code corresponding to the correct response. However, the code we showed wasn't correct in terms of what it printed for the `powered_by` field. We retrieved that value from the code that was already there, instead of from the LSP response that was received from the language server.

Fix it so that we print out the actual response retrieved by the language server.

Reviewed By: jewelpit

Differential Revision: D16656990

fbshipit-source-id: 95bb9b97eb6584f1287c4427253e05dbe2edc48a