A representative subset of errors that currently appear on declarations, but maybe shouldn't
Summary:
Currently many of our errors appear on class/interface/trait declarations, rather than the members that cause them. This made it easier to suppress bad classes in the past, but now that hack is more mature it's making it harder to categorize and programmatically solve the remaining issues.

This adds some tests so we can try to split that up confidently.

Reviewed By: DavidSnider

Differential Revision: D16612542

fbshipit-source-id: 29022008d1900323aa717e68eb6ce7551583268f