introduce preg_replace_with_count
Summary:
As we are on the final stages of removing references support from HHVM, we need an alternative solutions for builtin functions, that take optional reference parameters.

For now `preg_replace_with_count()` takes a required reference argument, but I will be switching it to inout soon, as well as removing $count parameter from `preg_match()`.

Reviewed By: billf

Differential Revision: D16664041

fbshipit-source-id: 7965a98aa99017929ca3150b572e1c759b447024