Allocate space for out parameters in conjureBeginInlining()
Summary:
When we compute the cost of inlining we reserve space for the caller's locals and iterators. Since num_locals is generally a superset of the number of inout parameters passed to any callee we hadn't seen any failures from this inlining code for inout functions (despite the out parameters technically overwriting locals in these translations). And since this code was never actually going to be run it was never an issue.

If, however, hhbbc is able to elide locals associated without parameters then the writes will happen after the end of the stack triggering assertions.

Reviewed By: ricklavoie, binliu19

Differential Revision: D16696361

fbshipit-source-id: a0e70b1b01760452519e478ba602b6ba7b7019df