introduce is_callable_with_name
Summary:
As we are on the final stages of removing references support from HHVM, we need an alternative solutions for builtin functions, that take optional reference parameters.

For now `is_callable_with_name()` takes a required reference argument, but I will be switching it to inout soon, as well as removing $callable_name parameter from `is_callable()`.

Reviewed By: paulbiss

Differential Revision: D16698295

fbshipit-source-id: 20a4c9b71e5645eeba15cd0f7134cc7a9f753ea1