Replace expr_annotation by a hint_annotation in type_hint
Summary:
The `type_hint` type (which is a `expr_annotation * hint option`) used to store the hint and the localized inferred type in `f_ret` was confusing as we were always setting the position of the `expr_annotation` part to `Pos.none`.
This diff adds a new annotation to the codebase, `hint_annotation` which is used to represent a type without position.

Reviewed By: andrewjkennedy

Differential Revision: D16494949

fbshipit-source-id: 612cba54e8ffc32dc8644ea95e4944fa02b1825f