Check expression is LHS before getting members
Summary:
Don't call `get_member` or `get_static_member` before checking that we're in the LHS of an assignment.

`get_member` and `get_static_member` have a side-effect of recording dependencies (`add_idep`). As a result, this TAST check would add `foo` as a property when it saw methods like `$x->foo()`.

Differential Revision: D16668006

fbshipit-source-id: 3c0ec6537c00cf4f4e183fb6439d4a03d018ed32