Add handling of Bucket.wait in shared_lru
Summary:
Clients of `shared_lru` can now return a `Bucket.wait` state if they want to wait for all pending jobs to complete.

Due to this change, some of the logic in `shared_lru.ml` was able to be simplified:

- `fork_and_start_workers` only starts a single worker, and was renamed appropriately.
- `monitor_workers_h` is responsible for starting all workers, and workers do not have to be "pre-initialized" with work beforehand.

I don't think that this refactor will have an effect on perf/memory usage, but I'm running tests in comparison with the current implementation, and against `merge-base`.

Update: [4 reps](https://fburl.com/everpaste/7v11nqqo) finished, and it looks like this implementation saves memory (and might be better in terms of performance as well, more reps are necessary) – but it does not appear to cause any regressions!

Reviewed By: 2BitSalute

Differential Revision: D16634639

fbshipit-source-id: 76caa3715bd46e5f21fb165395805f0413590b3a