Open-source serverless IDE
Summary:
The "serverless IDE" is a project to provide IDE services using the bare minimum amount of information necessary, and not depend on a heavyweight `hh_server` process being up and running. This is primarily useful at Facebook-scale, but could also improve the latency of common IDE operations for open-source users.

Usually, the serverless IDE requires a "naming table" to provide IDE services with low latency. This naming table is a symbol table which tells the process where all of the symbols in the codebase are. We download it using internal infrastructure, so it will fail in the open-source build. However, you can generate the naming table like this:

```
hh_server --check . --save-naming path/to/naming/table.sqlite3
```

Then launch your LSP client with `hh_client lsp --serverless-ide`, and pass the path to the naming table in the initialization options for the `initialize` request:

```
"initializationOptions": {
    "namingTableSavedStatePath": "path/to/naming/table.sqlite3"
},
```

The language client will then load the naming table and start providing some IDE services without requiring `hh_server` to be running. Note that currently it only provides IDE services if `hh_server` is known to be stopped.

Reviewed By: 2BitSalute

Differential Revision: D16697329

fbshipit-source-id: cf811565f1b14548028072dfd46431d3b9d8c2b3