Support `debug` method to issue `telemetry/rage`
Summary:
During debugging, it can be helpful to issue `telemetry/rage` to get details about the language server's internal state. See the documentation on the new method.

I tried to render the `debug` message's context, but ran into some issues and didn't want to spend more time debugging it (heh). In the course of debugging, I renamed all the line-number variables to be explicit about whether they're zero- or one-indexed, and whether ranges are inclusive or exclusive. I left that in this diff since it seems helpful in general.

Reviewed By: ljw1004

Differential Revision: D16701426

fbshipit-source-id: 94682c187d0ea1c809278d50dc070f1188219a54