Add support for writing to disk in LSP tests
Summary:
Add support for writing files to disk in the middle of LSP tests. Previously, it was difficult since you could write a file to disk either before or after the test had run, but not in the middle of it.

Some care has to be taken about synchronizing certain events, as discussed in the doc-comment for the new `write_to_disk` method.

This is useful for testing incremental processing in the serverless IDE.

Reviewed By: ljw1004

Differential Revision: D16701424

fbshipit-source-id: 3494ea42728f876290cf3a1a5e35f499be551cb6