Track provenance tags for arrays during constprop
Summary:
In support of HAM--this extends the notion of array provenance to static arrays in the repo.

We assign a (filename, line #) tag to every static array in the assembled unit from hackc and extend hhbbc's inference to propagate these tags around in the same way the runtime would so that any static arrays hhbbc adds to the repo are also tagged.

This doesn't yet apply to empty arrays (later in stack) but for now, we get by by just tacking the provenance info onto any arraylike data that we happen to come across, and then emit this along with the array datas back into the repo.

Reviewed By: mxw

Differential Revision: D14935850

fbshipit-source-id: 35a67b43e5ec3a0f0285b0231432201157b806d6