Set LRU cache size to 15MBs
Summary:
Wanted to test a suspicion that the size of the cache did affect perf, also Tatiana suggested testing cache evictions, and this led to some interesting results. Currently, with a 10GB cache, we have ~50% performance regression.

For reference, cachelib uses ~4MBs for metadata, and we were attempting to store 136MB of uncompressed data (everything in `decl_provider.ml` except classes).

This also implies that with compression, we could make the cache size even smaller!

All of the following results are with 8 reps, and the perf regression described is for full-inits:

- [Cache Size of 10MBs](https://fburl.com/everpaste/vci4tkdm) – Regression of 49%
- [Cache Size of 15MBs](https://fburl.com/everpaste/lz4attlm) – Regression of 22%
- [Cache Size of 20MBs](https://fburl.com/everpaste/q8550wwj) – Regression of 37%
- [Cache Size of 25MBs](https://fburl.com/everpaste/atvh0qz2) – Regression of 36%
- [Cache Size of 30MBs](https://fburl.com/everpaste/becme0kv) – Regression of 52%

Note that if **anything** else is stored in the cache, this number will need to be tweaked again to find the right balance.

Reviewed By: 2BitSalute

Differential Revision: D16712681

fbshipit-source-id: eeaa2aaee37d0d33331a67e94d7facac9f4333a7