Quick fix for decl errors in other files
Summary:
Earlier changes under the use_new_type_errors flag changed the file of some errors in addition to the position of them. That will eventually cause problems with some scaling options we're looking into. Plus it's a little squirrelly to blame a trait when in reality the trait only causes a problem in some/one of its users.

This rolls back that behavior when the more specific error would change the file of the reported error.

Reviewed By: Wilfred

Differential Revision: D16720778

fbshipit-source-id: 109b6c8f8698d4934dc63d3ec7b3ea1d5679dc91