Added test for behavior of use_new_type_errors flag across files
Summary:
The error repositioning under the `use_new_type_errors` flag sometimes causes the file that errors are reported in to cross file boundaries. That is not great as it limits our general approach to processing one file at a time.

This test shows the most common scenario, where a trait's implementation (in some other file) causes a class error.

Reviewed By: DavidSnider

Differential Revision: D16720265

fbshipit-source-id: cf8fe96906a683add25038ce7fb24d1733dc7630