Explicit attribute for generics
Summary:
Sometimes it's useful to insist that generic arguments to functions, methods, and classes are provided explicitly rather than inferred. This is especially true of signatures in which the type parameter appears only contravariantly (so Hack is free to infer `mixed`, or the upper bound on the type) or only covariantly (so Hack is free to infer `nothing`, or the lower bound on the type).

The attribute `<<__Explicit>>` serves this purpose. It can be used on any generic parameter. If any type parameters to a function, method, or class are marked in this way, then the use-site must provide explicit types in angle brackets (e.g. `foo<int,string>(...)`. Type arguments corresponding to parameters that are *not* marked `<<__Explicit>>` can be wildcards (`_`), indicating that Hack should infer them.

The redundant-generics lint rule now ignores any type parameters marked with `<<__Explicit>>`.

Reviewed By: kmeht

Differential Revision: D16754038

fbshipit-source-id: 213cfc3e3e5c14d9af88f1a41730f17abd27d473