Fix an unsoundness with calling a static method through the classname of a generic class
Summary:
The following example type-checks:

```
<?hh // strict
// Copyright 2004-present Facebook. All Rights Reserved.

interface I {
  public function f(): void;
}

class C<T as I> {
  public static function g(T $x): void {
    $x->f();
  }
}

function test(int $x): void {
  $c = C::class;
  $c::g($x);
}
```
but the call to `test(42)` fatals with
```
Fatal error: Uncaught exception 'BadMethodCallException' with message 'Call to a member function f() on a non-object (integer)' in /home/manzyuk/scratch/static_method_call_classname_with_generics.php:10
```
The problem is that the type of `C::class` is inferred to be `classname<C<#1>>` for a fresh type variables `#1`.  (Every class is considered to have a(n implicitly defined) constant called `class` of type `classname<this>`.)   But the constraint `#1 <: I` is not added to the environment, allowing the static method g to be called with an argument of any type.

Reviewed By: kmeht

Differential Revision: D16730612

fbshipit-source-id: c4223c4bd0da83070e7982a30099ba995a1f9fd1