Extend extraction of class name from Aast hint
Summary:
When generating code for a class, we specify the class it extends and interfaces it implements (if present).  To do so, we look up `Aast.class_.c_extends` and `Aast.class_.c_implements`, which are lists of hints. Thus we need to extract the name from Aast hint. In the basic version we supported the main case: `Happly`, but there are more: for example,
`class A extends B<this::ATypeConst>...`.

Reviewed By: andrewjkennedy

Differential Revision: D16667995

fbshipit-source-id: fcc3219d6c2d049d997d409c8bca0ffafcf52821