Do not look up constructor's origin
Summary:
We look up the origin of methods and properties (the class where the accessed method or property was defined) to know where to declare them. For example, in the following scenario:
```
class A {
  public function f(): void {}
}

class B: extends A{}

public function useB(B $arg): void {
  $arg->f();
}
```
we will record (among others) the dependency `Method(B, f)`, but we should declare `f` in its origin `A` and not in `B`.
However, the constructor's origin is obvious -- it's always the constructor's class! Also, this check used to raise `DependencyNotFound` when no constructor is declared and the default constructor is used.

Reviewed By: 2BitSalute

Differential Revision: D16647329

fbshipit-source-id: 68d2e2e40de37ea49475d6b7c24b0c09b2b27b5a