Do not convert polymorphic variants
Summary: As shiqicao noted in the summary of {D16751979}, it would be possible to generate correct interop code for polymorphic variants, but it would take a little more doing. If we have a need for them in the future, we could invest time in generating proper interop code for them, but since we don't need them now (as of {D16767537}), we just shouldn't generate Rust types for polymorphic variants at all.

Reviewed By: shiqicao, jewelpit

Differential Revision: D16767729

fbshipit-source-id: 52f2c33e05526cb30ca3e20ba3b4b4ecd5c72026