Remove polymorphic variant Typing_defs.dependent_type
Summary: As in D16751979, this polymorphic variant isn't strictly necessary--a regular variant type serves our needs just fine, and we don't yet have support for polymorphic variants in generated Rust interop. As noted in the summary of D16751979, it would be possible to generate interop code for polymorphic variants, but since we do not use them extensively (or in ways that are not expressible using regular variant types), it is easier to get rid of the few instances we do have.

Reviewed By: shiqicao

Differential Revision: D16767537

fbshipit-source-id: 7a812c2bcc9841524659541458301ba58bd12813