Server should respect max_errors
Summary:
`max_errors` is properly handled in `hh_single_type_check`, but it was ignored in everything but the header in Hack client/server.

This change makes it so the client passes `max_errors` to the server, which is going to return that number of errors and the count of dropped errors, if `max_errors` is set.

Reviewed By: jewelpit

Differential Revision: D16751990

fbshipit-source-id: 4d53e1120fbe49021ee77f9a5056dea7e69ad0a8