Allow dict to be used as a callable
Summary: PHP arrays of 2 elements with integer keys 0 and 1 can be passed to functions expecting callables or callbacks. This support was extended to vecs, which continue to be backed by a packed array type. Other Hack array kinds (dict, keyset) are not callable today. Long term, we plan to drop support for darrays (associative PHP arrays), but to do so we'll need them to behave like their replacement. Since darrays are callable, this diff aims to make dicts callable as well. Note that the introduction of function pointers to HHVM will eventually render it unnecessary to support invocation of arrays.

Reviewed By: jano

Differential Revision: D16653959

fbshipit-source-id: cce4e0e0ffc832b5cfbdeffbef10442380670a54