Ban visibility on type constants in the typechecker.
Summary: We recently added experimental support for visibility on type constants. It doesn't do anything; e.g. see: P78812083 and P78849487. We forgot to ban it while it's being worked on and it accidentally snuck into www. Ban them in the typechecker (gated by the same option that bans visibility on constants).

Reviewed By: periodic1236

Differential Revision: D16780932

fbshipit-source-id: 951552028b5f669e9bc5131b8068de328b81d1b8