Add "enforced" attribute to parameter and return types
Summary:
When `--coercion-from-dynamic` is set, we want to be able to pass values of type `dynamic` to functions whose parameters have types that are enforced at runtime, and return values if the return type is enforceable. Currently we plumb a lot of "decl" type information around in order to test its enforceability in `Typing_coercion.supports_coercion_from_dynamic`. However (a) enforceability depends only on the decl type, not on anything else, and (b) the current scheme doesn't naturally extend to the use of first-class functions through `inst_meth` and lambdas.

Instead, we can compute enforceability upfront, and build it in to function types. The plumbing of decl types is no longer necessary, and we get coercion from dynamic for first-class functions.

Reviewed By: vassilmladenov

Differential Revision: D16679060

fbshipit-source-id: 97de57a2f7a2281cd2d82b2072c3f3f1e14b1613