Create a global `Provider_context.t`
Summary:
All of the providers will be required to check this context first before calculating the value that they're looking for from disk or retrieving it from cache.

This helps us distinguish between values that are stored in memory part of the delta from the repo, and values that are cached in memory. We currently don't do this in a principled fashion, which can lead to state-corruption bugs.

Reviewed By: jewelpit

Differential Revision: D16725112

fbshipit-source-id: 051e1b543c501720dc1f33f41ab5f8efa0d7dac4