Split `ServerIdeContext` into `Provider_context` and `Provider_utils`
Summary:
The "server IDE context" logically belongs with the providers, as it provides a delta that can be applied to the repo state for use with the providers.

In particular, we access all the providers via global state, and that global state should check the delta first. Either we can pass the provider context all the way down the stack, or we can make it another piece of global state that the rest of the providers should check. We do the latter in the rest of this stack.

It's currently the case that some providers do *not* check the context that we're currently using, and that makes it possibly to occasionally mix up ASTs for different versions of the same file, which causes sIDE correctness bugs.

Reviewed By: jewelpit

Differential Revision: D16725111

fbshipit-source-id: 6eecb5034970563b4119495396ff0f05b11abd9f